add api for partitioner

+++ src/circuit/pin.h
@@ -24,7 +24,7 @@ class Pin {
 
  public:
   Pin(std::pair<const std::string, int>* name_num_pair_ptr, BlockType *blk_type);
-  Pin(std::pair<const std::string, int>* name_num_pair_ptr, BlockType *blk_type, int x_offset, int y_offset);
+  Pin(std::pair<const std::string, int>* name_num_pair_ptr, BlockType *blk_type, double x_offset, double y_offset);
 
   const std::string *Name() const {return &(name_num_pair_ptr_->first);}
   int Num() const {return name_num_pair_ptr_->second;}
